OMAP3: PM: Errata i539: McBSP clock management to allow CORE idle
Work around for Errata i539.  When enabling a McBSP clock belonging to
the PER domain, another functional clock in PER must be enabled to
allow CORE to idle.  GPIO2 functional clock is enabled when McBSP[2..4]
is requested, as it uses the 32 KHz clock (to minimize impact on power
consumption).

reffered patch: http://review.omapzoom.org/#/c/14029/

Change-Id: I4b251d8ab56c6673049ff4f38f1439fe096086d8
Signed-off-by: Hong H. Pham <hong.pham@ti.com>
Signed-off-by: Deepak Muddegowda <x0171695@ti.com>