OMAP3: PM: Workaround for errata i581
When internal SWAKEUP event occurs while CORE DPLL is going to bypass,
PRCM may report that DPLL3 is locked whereas it is not.  This is due to
the timing path of an internal signal which is not fitting to one L3
clock cycle.

In that case, PRCM DPLL control FSM deasserts SDRC_IDLEREQ signal before
DPLL3 Lock state is set.

Consequence is that SDRC is released from IDLE with bypass clock (which
is too low) instead of the locked frequency. DLL may/maynot lock base
on Process Voltage Temperature conditions

reffered patch : http://review.omapzoom.org/#/c/14171/

Change-Id: I28ed028b212c2cb513c2b8832514d4cc63451118
Signed-off-by: Steve Schefter <sschefter@ti.com>
Signed-off-by: Deepak Muddegowda <x0171695@ti.com>