OMAP3: USB-EHCI :Work Around for errata i464
Workaround for EHCI suspend-resume and suspend-remote-wakeup

We need to halt the EHCI controller before clearing the resume bit.
This allows the local and global framecounters to be synced up. The
controller is unhalted immediately after clearing the resume bit in
PORTSC_i.

This covers both suspend-resume and suspend-remote-wakeup cases

reffered patch : http://review.omapzoom.org/#/c/1293/1

Change-Id: I912d261ec9c9b7cf162dda7d96dbcb5d42ed78cd
Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Deepak Muddegowda <x0171695@ti.com>