ti-st: fix chip_asleep/awake & wake-ack everytime
chip_awake platform hook needs to be invoked from TI shared transport driver
when both the chip sends wake-indication intending to wake up and also when
host intends to wake up the chip.
Also, fix the function parameters sent across in these platform hooks,
although the parameters are for purely for future reference & currently not
made use of.

This patch also sends across wake acknowlegment even when there is a collision
of wake-indication, since this is pretty much ignored by the chip.
However, this is useful when UART for varying reasons drops the wake-ack &
chip retries the wake-indication.

Change-Id: I690247afb80eb1aca24f79feaaddce85d8473a05
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>
Signed-off-by: Andrey Gostev <fga022c@motorola.com>