Add ZRAM_FOR_ANDROID, with all of the bits
I am not 100% if this is a "must-have-feature" or not.
In fact i was able to reproduce the performance boost
with zram(10%) enabled, however i had backport a lot
of new kernel features in vmscan.c and swapfile.c.

Furthermore this commit includes a lot of bug fixes
for lmk, so only disable CONFIG_ZRAM_FOR_ANDROID.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>

Conflicts:
	drivers/staging/android/lowmemorykiller.c
	drivers/staging/zram/Kconfig
	include/linux/mm_types.h
	mm/swapfile.c
	mm/vmscan.c