Added base: genlock-support
Added genlock for kernel. From the documentation:
'genlock' is an in-kernel API and optional userspace interface for a generic
cross-process locking mechanism. The API is designed for situations where
multiple user space processes and/or kernel drivers need to coordinate access
to a shared resource, such as a graphics buffer.

Also it enables better compatibility for ICS+.
Has to be nebaled in defconfig: CONFIG_GENLOCK=y

Change-Id: I0af22c677b61bc2ab9ed01eb33e7ca63e7fe314b
Signed-off-by: Blechd0se <alex.christ@hotmail.de>