cpufreq: interactive: integrate some new features
Since we use interactive as our default governor on
JB Roms, we should update it more frequently.

This patch introduces some new featues such as io_is_busy
which helps to improve run-time of IO boud tasks, this also
introduces some necessary kernel/timer changes for this to
work.
On every tap there is boost pulse and now the duration
of it is tunable from the userspace.
Finally there is now a targe_load attribute to specify
how aggressiv interactive should adjust speed to meet
the observed load.

Change-Id: I9eb356db1d2701caed4f22ca3c13a86f34df8642
Signed-off-by: Blechd0se <alex.christ@hotmail.de>