cpufreq: interactive: add suspend/resume logic
If the display of the phone is off, it won't need
high frequencies probably, so we can limit it to the
lowest possible frequency. After resume (display on)
we can simply restore the old frequency which was used
before.

This should not only increase battery life a bit, but
also can improve performance in some workloads.

Also some tunables were slightly adjusted.

Change-Id: I27a18c23238001f398a8eb7c01597b2f070e578a
Signed-off-by: Blechd0se <alex.christ@hotmail.de>