zcache: fix kernel panic
Before this patch, the kernel panics nearly on every
tick as long as zcache is enabled. So i disabled
all ASSERT_SPINLOCK() calls, because there is probably
some code missing for lockdep_assert_held().

ZCache is now somewhere between Kernel 3.2 and 3.3,
Cleancache is on 3.4.

Further updates will give the opportunity to decide
which compression methode should be used for zcache (lzo for now).

To enable it; simply add "zcache" in cmdline, to check if its
working look at /sys/kernel/mm/zcache

Change-Id: Iac6b9a156f22cd7803e4daf0d65fdb2681e2599f
Signed-off-by: Blechd0se <alex.christ@hotmail.de>