gpu: add basic gpu control driver
This patch series will introduce a new gpu driver
which was written completely from scratch.

I noticed that our gpu stays all the time on
its max frequency (200mhz). This is not really
efficient and probably costs some serious battery.

This frist patch will enable the well know earlysuspend
handling. If you turn off your screen it will set
the freq to 50mhz to save some battery.

Also with this patch it is possible (in theory) to
overclock/underclock your gpu on-the-fly, however
this needs much more work. The goal is to implement
a lightwight gpu scheduler which sets the frequency
automatically and let the user decide what the max_freq
should be.

Change-Id: I38d441cf44d8a4994f0d2cc8d45cd9b2a6949169
Signed-off-by: Blechd0se <alex.christ@hotmail.de>