cpufreq: interactive: add unregister_early_suspend
If suspend_enabled is set to 1 and the device screen goes off,
the device can't change the governor anymore.
This is probably due to the suspend/resume code, because with
suspend_enabled set to 0 it won't happen.

So the first idea is to unregister the suspend handler
once the governor exits.

Change-Id: I25b8cfedeb06a58cca25df4c42a151cb106763e5
Signed-off-by: Blechd0se <alex.christ@hotmail.de>