cpufreq: interactive: add load based suspend
Some users reported that some background task won't
work as they worked before the new suspend/resume logic.

The easiest way would be to bump up suspend_freq to 600 Mhz,
but this would result in wasted cpu power and is also
not efficient. This patch tries to address the issue
with two approaches;

1.) The simple one
    Instead of boosting the cpu MAYBE to max_freq on
    resume, boost it in any case to 1 Ghz.

2.) The load balanced one
    Like in one of the previous implementation i added
    a cpu_load value. If we are suspended and above 85%
    cpu_load we scale to the next available frequency.

In my test scenario this ensured smooth audio-playback
as well as a fast wake up. Any wakeup delay should be now
in the hands of Android itself.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>