cpufreq: ondemand: add input_boost
Project Butter introduced the boost logic for
interactive. Ondemand can do the same, with a new
sysfs called input_boost. Default is 0 (on).
This will boost the cpu on each input from the
touchscreen to the max_freq available.

Ondemand should be now a serious alternative
compared to interactive for android devices.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>