clk: add gpio controlled clock
The added clk-gpio is a basic clock that can be enabled and disabled
trough a gpio output. The DT binding document for the clock is also
added. For EPROBE_DEFER handling the registering of the clock has to
be delayed until of_clk_get() call time.

Fixed return value from clk_register_gpio.

Change-Id: Iafce1c690626b3ee0c6e2835c15f6d14f63e00a2
Signed-off-by: Jyri Sarha <jsarha@ti.com>
CC: t-kristo@ti.com
Signed-off-by: Jim Wylder <jwylder1@motorola.com>