omap2: allow to disable margin for OPP
In order to fix overclocking we need to be able
disable the margin for a specific OPP. This
patch takes care of this.

In my testcase i added a frequency (1100 mhz) and locked
voltage for this specific frequency at 1.375V.

Anyway, the "big" problem is 1.2Ghz, since it doesn't
even run with 1.5V stable for me.

I also increased the margin vor the other frequencies,
so if smartreflex feels in the right mood, it
can set lower voltages on the fly.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>