freezer: fix sleep of death with binder
When going from idle->suspend (without usb plugged in)
our pvr driver crashes because of a bad freezer call.
After pvr all process died and the phone reboots with a
hot reboot.

This was caused by binder using the new freezer code.
This fixes the sleep of death and also keeps functionality
for binder.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>