isl29030: fix reboot of proximity sensor
The proximity sensor would crash immediatly if
there was a out- or incoming call. Thanks to the
working /proc/last_kmsg this issue is fixed. The
input_handler for EV_ABS events crashed because it
used a newer api level.

Our moto driver requires this old 2.6.32 legacy
code. If any problems come up with this patch,
we need to set a seperated flag for EV_ABS events.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>