IKG5-1476 dts: minnow: Support multiple device tree revisions
Create hardware revision based dts files that include the common
configuration from omap3-minnow.dts.  The property mmi,hwrev
will be used by the bootloader to identify the matching device tree.

casper-p1 is a legacy board that predates the first real
minnow board.  As hardware differences are identified, they
should be migrated from the omap3-minnow.dts to the appropriate
revision based dts file.  Or if the majority of revisions are
expected to be configured identically, the setting can remain
common and overridden as needed.

Change-Id: I328044668f51d000612cfe0541bc464fcec6fe3d
Signed-off-by: James Wylder <jwylder@motorola.com>