smartreflex: turn recalibration into dedicated wq
Before this patch the recalibration was just like the
normal voltage calibration a normal scheduled work, which
means it has some sort of sharing pool with other workqueues.

In my research i found, that the normal calibration process
was rather sooner than later and that means, the recalibration
work is the only after some minutes of uptime.
This turns the work into a dedicated delayed work with the exact
same functionality to reduce the overhead and let it calibrate
quicker. Also, adding earlysuspend handler to stop work
during suspend, because its very unlikely that we want to calibrate
again when we are near deepsleep.

All in all this *should* reduce battery consumption a bit,
the next step is to teach smartreflex to stop even the calibration
process if we are really suspended and in deepsleep if it doesn't
know about this.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>