cpufreq: interactive: lets change the suspend code again
Instead of purely locking the freq to a certain value
we can just add some sort of limit for the suspend-code.
This means interactive suspends with 600Mhz and during the
deepsleep race the governor might to decide to go down to
300Mhz.
However, if there is any load during that time, the governor
might want to jump up again. Instead of letting it do so,
it only scales up to 600Mhz when the target_freq is in range
(meaning between 300 and 600 on a default setup).

Also deleting some deprecated stuff here and changing a
few tunables.
Lets wait for angry nightly reports :)

Signed-off-by: Blechd0se <alex.christ@hotmail.de>