IKXCLOCK-412 mmc: HACK: mmc_claim_host leak in mmc_blk_issue_rq
The function mmc_blk_issue_rq will claim the mmc_host multiple
times in "some cases" causing the host claim count to never
return to zero.  The claim in mmc_blk_issue_rq includes a
comment "claim host only for the first request." Further, when
all the requests have been serviced, only one release will
ever be called.

Here additional claims are prevented by using a boolean host_claimed.
This is sufficient if the function is fully serialized.  The nature
of the existing code suggests that is expected, but the existence of
this behavior suggests otherwise.

Change-Id: I6c28eb9810baaf6d1d303553bcd33c8e62a424f5
Signed-off-by: James Wylder <jwylder@motorola.com>