IKXCLOCK-583 arm: omap2: pm should set suspend in early resume
On omap34xx, the suspend flag in pcrm sets a clear of the irq
register so that interrupts are deferred during early stages of
resume.  This code must not be called prior to suspend or the device
will not wakeup for the pcrm domain.

For omap34xx devices, the PRM MPU block apparently must be masked
during early stages of resume. This is currently done by setting the
suspend flag in omap_pm_begin(). This should then be read during the
interrupt handler during wakeup and used to save and clear the pending
interrupt. The interrupt is then disabled. Once enough hw is enabled,
the interrupts are restored and the interrupts are re-enabled.

In cases where the device is not resumed from the MPU block, pending
interrupts can fire during the suspend phase, causing the IRQs to be
disabled during suspend.  The setting of the suspend flag is moved to
early in the resume phase to prevent any disabling of interrupts during
suspend.

Change-Id: Ia09664b89af41bd5879620722624bf3169d9eff4
Signed-off-by: James Wylder <jwylder@motorola.com>