IKXCLOCK-639 tps65912: irq handler accesses spi bus too early
Late is suspend and early in resume, the irq handler for
the tps69512 driver will attempt to access the spi bus to
determine the origin of the interrupt.  This may happen
before the spi bus is resumed.

Force the handler to delay until the associated spi
device has been resumed.

Change-Id: Ie38185490d8da9555c3092f543b0c495d1de5782
Signed-off-by: Jim Wylder <jwylder1@motorola.com>