IKXCLOCK-714: Change download_status permissions
Change-Id: Ieb2df00e151d3f0b0f9522bab1739af2f468123e

Change the permissions of download_status sysfs to global. Sensorhal waits
on download_status sysfs to know if m4 is ready to handle sensor requests.
But due to the permissions being set to IRUSR, until sensorhal makes changes to
deal with permission changes for download_status, kernel needs to handle the
scenario