IKXCLOCK-781: Temporary disabling C55 UART connection to OMAP
There are problems reported where xClock lockups during normal usage.
By trial and error, it has been identified that disabling UART1
on OMAP fixes the lockup. Pushing this in the baseline to get it
out to a substantial number of users to see if C55 interactions
really are causing the lockups.

Change-Id: I13cb2940529853a470381cc89f70c681d5551484