IKXCLOCK-807 kernel: send KEY_SLEEP when palm is pressed
clockworks requires kernel to send KEY_SLEEP when palm is
is detected on ITO.  In ATMEL this is done by using key suprression
feature which is defined in T42.  Once enabled, send KEY_SLEEP 1 then
send KEY_SLEEP 0 when supression is deactivated.  This is done via
firmware and there is no software invloved.

Change-Id: Ib2c33d04fb2ed46bc5309af28009d743e206ed5a