Disable unneeded Bluetooth kernel config settings
After switching from BlueZ to Bluedroid we need only the following
kernel config settings to be enabled: CONFIG_TI_ST and CONFIG_ST_HCI.
After building kernel config it appears that there are few other changes
in defconfig that are required in order to keep kernel configuration in
a consistent state. This commit will address that issue as well.

Change-Id: If07fa5141240e85d101ad2396b7db3fe51a34e3b
Signed-off-by: Andrey Gostev <fga022c@motorola.com>