IKXCLOCK-1044 gpio: omap: enable OMAP offmode
Output GPIO lines do not stay in correct state after exit
from OFF mode, it is happening because GPIO modules does
not keep state of registers if module is not in wake up
domain. Work around is to save pad and gpio configuration
before entering OFF state and put pads to safe mode with
pulls enabled according with level on GPIO output
After exit OFF state all pad configurations are restored.
This work around is related to GPIOs which are configured
as output.

Change-Id: Ia0ca337de1bda28dd5f01c9d51686d822fc9ec21
Signed-off-by: Vladimir Tsunaev <vladimirt@motorola.com>