IKXCLOCK-2527 watchdog: omap_wdt: change pm_runtime to IRQ safe
WD suspend function is using pm run-time API to suspend, and pm
run-time for WD is initialized with no IRQ safe, as result IRQ will be
enabled for some period of time to handle pm run-time request for WD.
If there is any pending/firing IRQ during this time it will be handled
and if IRQ is threaded then the thread will be put into running queue,
but SW is in idling sequence (no scheduling) and at this point there
is nothing to prevent system go to sleep (IRQ is not pending , it was
handled), as result the IRQ thread will be scheduled only after some
wake up event with unpredictable delay.

The solution is make WD pm runtime IRQ safe, so if there is pending
IRQ, system just will not go to sleep.

Change-Id: Ieed3a01ce503fcd11c54053b18780f6081b3f039
Signed-off-by: Mohsan Habibi <mohsan@motorola.com>