IKXCLOCK-1806 pvr ver 3: Disabling LLS (Low Latency Scheduling) for SGX DDK
Low latency scheduling divides each 3D task into 4 subtasks, where
each subtask renders a quadrant of the output buffer. This adds
additional overhead on both the CPU and GPU, and is not necessary
when working with the small output buffers we use due to the low
resolution on xclock's display.

SUPPORT_SGX_PRIORITY_SCHEDULING was not explicitly enabled before,
but was enabled automatically by SUPPORT_SGX_LOW_LATENCY_SCHEDULING.
Enable this feature explicitly so that it stays on when LLS is disabled.

This change must be made to the user and kernel drivers at the same time
or the GPU driver will not start.

Change-Id: I21081f37edc42e050a048ce864c5aa7d4552a69e