IKXCLOCK-2689 i2c: omap: reduce timeout and don't wait after transfer complete
Such a long timeout of 1s for a transfer to finish or wait for the
bus to become free causes an i2c client to become blocked for at
least 2s, if the attached slave is misbehaving. This timeout can
be reduced to 500ms; 500ms is plenty long enough for a single i2c
transfer to finish even in clock stretching cases.

Also, after an i2c transfer is complete, there's no benefit in
waiting for the bus to become ready, since the check will be done
again just before the next transfer.

Finally, in case the bus never gets freed, let's return an error
immediately after clearing the bus, regardless if the recovery
was successful. This will allow the i2c client to try the transfer
again on its own terms, instead of having the i2c driver attempt
to proceed with the transfer immediately after the recovery.

Change-Id: I6bc950948bd012aae4c3029b9c00603354a938fa
Signed-off-by: Mohsan Habibi <mohsan@motorola.com>