IKXCLOCK-2861 uart: Allow UART to grab timed wakelocks
In the case of asynchronous wakeups coming in from UART, the system
may need to stay awake for a period in order to catch a retry reliably.
By making a port specific platform data wakelock timeout parameter, a
port can be configured to stay awake longer than the retry period.

Change-Id: I46ab72bf19d9916c409cac9cdda51acc362fa1a9